+++ vdagent/vdagent.cpp
@@ -28,7 +28,6 @@
 #include <queue>
 #include <set>
 #include <vector>
-#include <cinttypes>
 
 #define VD_AGENT_LOG_PATH       TEXT("%svdagent.log")
 #define VD_AGENT_WINCLASS_NAME  TEXT("VDAGENT")
@@ -1102,7 +1101,7 @@ bool VDAgent::handle_clipboard_request(const VDAgentClipboardRequest* clipboard_
     UINT format;
     HANDLE clip_data;
     uint8_t* new_data = NULL;
-    intptr_t new_size = 0;
+    long new_size = 0;
     size_t len = 0;
     VDAgentClipboard* clipboard = NULL;
 
@@ -1144,7 +1143,7 @@ bool VDAgent::handle_clipboard_request(const VDAgentClipboardRequest* clipboard_
         goto handle_clipboard_request_fail;
     }
     if ((_max_clipboard != -1) && (new_size > _max_clipboard)) {
-        vd_printf("clipboard is too large (%" PRIdPTR " > %d), discarding",
+        vd_printf("clipboard is too large (%ld > %d), discarding",
                   new_size, _max_clipboard);
         goto handle_clipboard_request_fail;
     }
